Clean up and document various code.

+++ Bas.ADayAtTheRaces/Color.cs
@@ -11,13 +11,19 @@ namespace Bas.ADayAtTheRaces
     public sealed class Color
     {
         [DataMember]
-        public float Red { get; set; }
+        public byte Red { get; set; }
         [DataMember]
-        public float Green { get; set; }
+        public byte Green { get; set; }
         [DataMember]
-        public float Blue { get; set; }
+        public byte Blue { get; set; }
 
-        public Color(float red, float green, float blue)
+        /// <summary>
+        /// Constructs the color with the provided <paramref name="red"/>, <paramref name="green"/> and <paramref name="blue"/> values.
+        /// </summary>
+        /// <param name="red">The value of the red component.</param>
+        /// <param name="green">The value of the green component.</param>
+        /// <param name="blue">The value of the blue component</param>
+        public Color(byte red, byte green, byte blue)
         {
             Red = red;
             Green = green;
